+++ .changeset/metal-snakes-deliver.md
@@ -1,0 +1,6 @@
+---
+"ledger-live-desktop": patch
+"live-mobile": patch
+---
+
+New deeplinks + refactoring of deeplinks wrapper in LLM
