+++ apps/ledger-live-desktop/src/renderer/hooks/useDeeplinking.js
@@ -238,6 +238,10 @@ export function useDeepLinkHandler() {
 
           break;
         }
+        case "market":
+          navigate(`/market`);
+          break;
+
         case "portfolio":
         default:
           navigate("/");
